feat: Implementación completa de API REST para Aiden's Adventure
- ✅ Corregidos errores de compilación (nombres de archivos y clases)
- ✅ Implementada arquitectura completa: Repositorios, Servicios, Controladores
- ✅ Añadidas validaciones a todas las entidades
- ✅ Configurada base de datos H2 con consola web
- ✅ Implementados tests unitarios y de integración
- ✅ Creados endpoints REST completos para objetos y habilidades
- ✅ Añadido controlador de información del juego
- ✅ Implementado cargador de datos de ejemplo
- ✅ Actualizada documentación con ejemplos de uso
- ✅ Cambiado puerto a 8081 para evitar conflictos

Endpoints disponibles:
- GET /api/objects - CRUD completo de objetos
- GET /api/abilities - CRUD completo de habilidades
- GET /api/game/info - Información del juego
- GET /api/game/levels - Información de niveles
- Consola H2: http://localhost:8081/h2-console

+++ src/main/java/org/factoriaf5/projectohalloween/model/Enchanted_Sword.java
@@ -18,7 +18,7 @@ public class Enchanted_Sword {
     @GeneratedValue(strategy = GenerationType.IDENTITY) 
     @Column(name = "id_EnchantedSword")
 
-    private int id;
+    private Long id;
     private String name;
     private String type;
     private String efect;
@@ -30,7 +30,7 @@ public class Enchanted_Sword {
     public Enchanted_Sword() {
     }
 
-    public Enchanted_Sword(int id, String name, String type, String efect, Object object) {
+    public Enchanted_Sword(Long id, String name, String type, String efect, Object object) {
         this.id = id;
         this.name = name;
         this.type = type;
@@ -38,11 +38,11 @@ public class Enchanted_Sword {
         this.object = object;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
