+++ src/main/java/org/factoriaf5/projectohalloween/model/Health_Potion.java
@@ -18,7 +18,7 @@ public class Health_Potion {
     @GeneratedValue(strategy = GenerationType.IDENTITY) 
     @Column(name = "id_HealthPotion")
 
-    private int id;
+    private Long id;
     private String name;
     private String type;
     private String efect;
@@ -31,7 +31,7 @@ public class Health_Potion {
     public Health_Potion() {
     }
 
-    public Health_Potion(int id, String name, String type, String efect, Object object) {
+    public Health_Potion(Long id, String name, String type, String efect, Object object) {
         this.id = id;
         this.name = name;
         this.type = type;
@@ -39,11 +39,11 @@ public class Health_Potion {
         this.object = object;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
